+++ resources/source.rb
@@ -23,18 +23,18 @@ default_action :create
 attribute :name, name_attribute: true, kind_of: String, required: true
 
 # global parameters
-attribute :input_module, kind_of: String, default: 'im_file', required: true,
+attribute :input_module, kind_of: String, required: true,
                          equal_to: %w(im_exec im_file im_internal im_kernel
                                       im_mark im_mseventlog im_msvistalog
                                       im_null im_ssl im_tcp im_udp im_uds)
 attribute :destination, kind_of: [String, Array], default: ScNxLog::DEFAULTS_KEY
 attribute :input_type, kind_of: String
 attribute :exec, kind_of: String
-attribute :flow_control, kind_of: [TrueClass, FalseClass]
+attribute :flow_control, kind_of: [true, false]
 
 # not quite global, but somewhat common
-attribute :save_pos, kind_of: [TrueClass, FalseClass]
-attribute :read_from_last, kind_of: [TrueClass, FalseClass]
+attribute :save_pos, kind_of: [true, false]
+attribute :read_from_last, kind_of: [true, false]
 
 # im_dbi
 # this is undocumented as far as I can tell, so I'm not sure what to do here
@@ -42,13 +42,13 @@ attribute :read_from_last, kind_of: [TrueClass, FalseClass]
 # im_exec
 attribute :command, kind_of: String # required
 attribute :args, kind_of: Array
-attribute :restart, kind_of: [TrueClass, FalseClass]
+attribute :restart, kind_of: [true, false]
 
 # im_file
 attribute :file, kind_of: String # required
-attribute :recursive, kind_of: [TrueClass, FalseClass]
-attribute :rename_check, kind_of: [TrueClass, FalseClass]
-attribute :close_when_idle, kind_of: [TrueClass, FalseClass]
+attribute :recursive, kind_of: [true, false]
+attribute :rename_check, kind_of: [true, false]
+attribute :close_when_idle, kind_of: [true, false]
 attribute :poll_interval, kind_of: Float
 attribute :dir_check_interval, kind_of: Float
 attribute :active_files, kind_of: Integer
@@ -63,7 +63,7 @@ attribute :mark_interval, kind_of: Integer
 
 # im_mseventlog
 attribute :sources, kind_of: Array
-attribute :utf8, kind_of: [TrueClass, FalseClass]
+attribute :utf8, kind_of: [true, false]
 
 # im_msvistalog
 attribute :query, kind_of: String
@@ -85,7 +85,7 @@ attribute :ca_file, kind_of: String
 attribute :ca_dir, kind_of: String
 attribute :crl_file, kind_of: String
 attribute :crl_dir, kind_of: String
-attribute :allow_untrusted, kind_of: [TrueClass, FalseClass]
+attribute :allow_untrusted, kind_of: [true, false]
 
 # im_udp
 attribute :sock_buf_size, kind_of: Integer