+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Removed unnecessary default values from the source and destination resources.
+- Removed unnecessary name property from the papertrail resource
+- Use true/false as types in resource attributes
+
 ### Deprecated
 
 ### Removed